The code now works with autoclassifications on simple logic like XOR. It still does not learn properly with the given dataset.
Updated the sstem to use the types of input messages specified in the doc. Will add a readme/how to run for this project that explains that for Ethan.

Issues: It learns really well for simple data, but not complex data. I tried more hidden nodes, but not more than like 10 to avoid overfitting the data, still no dice.