Added almost everything neccecary for reading everything in from the files and stuff. Note that the current code does not contend with the fact that an item can be in multiple partnership relatipnships. See supplied data file "input24.txt", and look at binary simultanious relationships for the item C.
I don't really care how you want to fix it, and I think you know better how you want to fix it for the sake of how backtracking needs to change. So, change things however you like in the Item class to compensate for this, and then figure out how backtracking needs to be modified for it.

I'm confident I can figure out how to get the data for you there once you know how you want it formatted, an array of arrays, a dictionary, a pair of arrays, one showing partners and the other showing touples of bag pairings... Whatever you like.

All other data is read in correctly and creates pointers to the desired objects. I did a few tests with different files for different cases but there might be more cases I didn't think of. One I caught was the fact that an item can specify it needs to be in a or b, and then specify that it CANNOT be in b. Seems like a silly case but I guess it makes sense.

I'm off to bed now, idk when you'll see my commit and stuff but good luck, see you in class.